# expensetrackerapi/swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "title": "Expense Tracker",
    "description": "Expense Tracker",
    "version": "1.0.0"
  },
  "host": "localhost:1337",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/user/save": {
      "post": {
        "tags": [
          "User"
        ],
        "summary": "User save service",
        "description": "",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "description": "Request body.",
            "schema": {
              "$ref": "#/definitions/userSave"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User created successfully."
          }
        }
      }
    },
...
    "/organizations/retrieve": {
      "get": {
        "tags": [
          "User"
        ],
        "summary": "organizations retrieve service",
        "description": "",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "user",
            "in": "query",
            "description": "User.",
            "required": true,
            "type": "string"
          },
          {
            "name": "objectCode",
            "in": "query",
            "description": "Object Code.",
            "required": true,
            "type": "string"
          },
          {
            "name": "skipACL",
            "in": "query",
            "description": "skip ACL.",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "User retrieved successfully."
          }
        }
      }
    },
    "/organization/vieworganizations": {
      "get": {
        "tags": [
          "User"
        ],
        "summary": "organization vieworganizations service",
        "description": "",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "user",
            "in": "query",
            "description": "User.",
            "required": true,
            "type": "string"
          },
          {
            "name": "objectCode",
            "in": "query",
            "description": "Object Code.",
            "required": true,
            "type": "string"
          },
          {
            "name": "skipACL",
            "in": "query",
            "description": "skip ACL.",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "User retrieved successfully."
          }
        }
      }
    },
...
  }
}